+++ boards/Board.h
@@ -87,8 +87,8 @@ typedef struct
 	uint32_t (*RSAP)                (uint8_t type, uint8_t motor, int32_t value);   // restore axis parameter -> TMCL conformance
 	uint32_t (*SIO)                 (uint8_t type, uint8_t motor, int32_t value);
 	uint32_t (*GIO)                 (uint8_t type, uint8_t motor, int32_t *value);
-	void (*readRegister)          (uint8_t motor, uint8_t address, int32_t *value);  // Motor needed since some chips utilize it as a switch between low and high values
-	void (*writeRegister)         (uint8_t motor, uint8_t address, int32_t value);   // Motor needed since some chips utilize it as a switch between low and high values
+	void (*readRegister)          (uint8_t motor, uint16_t address, int32_t *value);  // Motor needed since some chips utilize it as a switch between low and high values
+	void (*writeRegister)         (uint8_t motor, uint16_t address, int32_t value);   // Motor needed since some chips utilize it as a switch between low and high values
 	uint32_t (*getMeasuredSpeed)    (uint8_t motor, int32_t *value);
 	uint32_t (*userFunction)        (uint8_t type, uint8_t motor, int32_t *value);
