+++ boards/TMC2100_eval.c
@@ -35,8 +35,8 @@ static uint32_t moveTo(uint8_t motor, int32_t position);
 static uint32_t moveBy(uint8_t motor, int32_t *ticks);
 static uint32_t GAP(uint8_t type, uint8_t motor, int32_t *value);
 static uint32_t SAP(uint8_t type, uint8_t motor, int32_t value);
-static void readRegister(uint8_t motor, uint8_t address, int32_t *value);
-static void writeRegister(uint8_t motor, uint8_t address, int32_t value);
+static void readRegister(uint8_t motor, uint16_t address, int32_t *value);
+static void writeRegister(uint8_t motor, uint16_t address, int32_t value);
 
 static void periodicJob(uint32_t tick);
 //static void checkErrors	(uint32_t tick);
@@ -297,7 +297,7 @@ static uint32_t GAP(uint8_t type, uint8_t motor, int32_t *value)
 	return handleParameter(READ, motor, type, value);
 }
 
-static void writeRegister(uint8_t motor, uint8_t address, int32_t value)
+static void writeRegister(uint8_t motor, uint16_t address, int32_t value)
 {
 	UNUSED(address);
 	UNUSED(motor);
@@ -305,7 +305,7 @@ static void writeRegister(uint8_t motor, uint8_t address, int32_t value)
 	TMCRhinoSA.setInt(value);
 }
 
-static void readRegister(uint8_t motor, uint8_t address, int32_t *value)
+static void readRegister(uint8_t motor, uint16_t address, int32_t *value)
 {
 	UNUSED(address);
 	UNUSED(motor);
