+++ boards/TMC2160_eval.c
@@ -34,8 +34,8 @@ static uint32_t GAP(uint8_t type, uint8_t motor, int32_t *value);
 static uint32_t getLimit(AxisParameterLimit limit, uint8_t type, uint8_t motor, int32_t *value);
 static uint32_t getMin(uint8_t type, uint8_t motor, int32_t *value);
 static uint32_t getMax(uint8_t type, uint8_t motor, int32_t *value);
-static void writeRegister(uint8_t motor, uint8_t address, int32_t value);
-static void readRegister(uint8_t motor, uint8_t address, int32_t *value);
+static void writeRegister(uint8_t motor, uint16_t address, int32_t value);
+static void readRegister(uint8_t motor, uint16_t address, int32_t *value);
 static void periodicJob(uint32_t tick);
 static uint32_t userFunction(uint8_t type, uint8_t motor, int32_t *value);
 static uint32_t getMeasuredSpeed(uint8_t motor, int32_t *value);
@@ -668,14 +668,14 @@ static uint32_t getMax(uint8_t type, uint8_t motor, int32_t *value)
 	return getLimit(LIMIT_MAX, type, motor, value);
 }
 
-static void writeRegister(uint8_t motor, uint8_t address, int32_t value)
+static void writeRegister(uint8_t motor, uint16_t address, int32_t value)
 {
-	tmc2160_writeInt(motorToIC(motor), address, value);
+	tmc2160_writeInt(motorToIC(motor), (uint8_t) address, value);
 }
 
-static void readRegister(uint8_t motor, uint8_t address, int32_t *value)
+static void readRegister(uint8_t motor, uint16_t address, int32_t *value)
 {
-	*value = tmc2160_readInt(motorToIC(motor), address);
+	*value = tmc2160_readInt(motorToIC(motor), (uint8_t) address);
 }
 
 static void periodicJob(uint32_t tick)
