+++ boards/TMC2208_eval.c
@@ -105,14 +105,14 @@ uint8_t tmc2208_CRC8(uint8_t *data, size_t length)
 }
 // <= CRC wrapper
 
-void tmc2208_writeRegister(uint8_t motor, uint8_t address, int32_t value)
+void tmc2208_writeRegister(uint8_t motor, uint16_t address, int32_t value)
 {
-	tmc2208_writeInt(motorToIC(motor), address, value);
+	tmc2208_writeInt(motorToIC(motor), (uint8_t) address, value);
 }
 
-void tmc2208_readRegister(uint8_t motor, uint8_t address, int32_t *value)
+void tmc2208_readRegister(uint8_t motor, uint16_t address, int32_t *value)
 {
-	*value = tmc2208_readInt(motorToIC(motor), address);
+	*value = tmc2208_readInt(motorToIC(motor), (uint8_t) address);
 }
 
 static uint32_t rotate(uint8_t motor, int32_t velocity)
