+++ boards/TMC2209_eval.c
@@ -109,15 +109,15 @@ uint8_t tmc2209_CRC8(uint8_t *data, size_t length)
 }
 // <= CRC wrapper
 
-void tmc2209_writeRegister(uint8_t motor, uint8_t address, int32_t value)
+void tmc2209_writeRegister(uint8_t motor, uint16_t address, int32_t value)
 {
-	tmc2209_writeInt(motorToIC(motor), address, value);
+	tmc2209_writeInt(motorToIC(motor), (uint8_t) address, value);
 
 }
 
-void tmc2209_readRegister(uint8_t motor, uint8_t address, int32_t *value)
+void tmc2209_readRegister(uint8_t motor, uint16_t address, int32_t *value)
 {
-	*value = tmc2209_readInt(motorToIC(motor), address);
+	*value = tmc2209_readInt(motorToIC(motor), (uint8_t) address);
 }
 
 static uint32_t rotate(uint8_t motor, int32_t velocity)
