+++ boards/TMC2210_eval.c
@@ -31,8 +31,8 @@ static uint32_t moveTo(uint8_t motor, int32_t position);
 static uint32_t moveBy(uint8_t motor, int32_t *ticks);
 static uint32_t GAP(uint8_t type, uint8_t motor, int32_t *value);
 static uint32_t SAP(uint8_t type, uint8_t motor, int32_t value);
-static void readRegister(uint8_t motor, uint8_t address, int32_t *value);
-static void writeRegister(uint8_t motor, uint8_t address, int32_t value);
+static void readRegister(uint8_t motor, uint16_t address, int32_t *value);
+static void writeRegister(uint8_t motor, uint16_t address, int32_t value);
 
 static void periodicJob(uint32_t tick);
 //static void checkErrors	(uint32_t tick);
@@ -421,14 +421,14 @@ static uint32_t GAP(uint8_t type, uint8_t motor, int32_t *value)
 	return handleParameter(READ, motor, type, value);
 }
 
-static void writeRegister(uint8_t motor, uint8_t address, int32_t value)
+static void writeRegister(uint8_t motor, uint16_t address, int32_t value)
 {
 	UNUSED(address);
 	UNUSED(motor);
 	UNUSED(value);
 }
 
-static void readRegister(uint8_t motor, uint8_t address, int32_t *value)
+static void readRegister(uint8_t motor, uint16_t address, int32_t *value)
 {
 	UNUSED(address);
 	UNUSED(motor);
