+++ boards/TMC2225_eval.c
@@ -105,14 +105,14 @@ uint8_t tmc2225_CRC8(uint8_t *data, size_t length)
 }
 // <= CRC wrapper
 
-void tmc2225_writeRegister(uint8_t motor, uint8_t address, int32_t value)
+void tmc2225_writeRegister(uint8_t motor, uint16_t address, int32_t value)
 {
-	tmc2225_writeInt(motorToIC(motor), address, value);
+	tmc2225_writeInt(motorToIC(motor), (uint8_t) address, value);
 }
 
-void tmc2225_readRegister(uint8_t motor, uint8_t address, int32_t *value)
+void tmc2225_readRegister(uint8_t motor, uint16_t address, int32_t *value)
 {
-	*value = tmc2225_readInt(motorToIC(motor), address);
+	*value = tmc2225_readInt(motorToIC(motor), (uint8_t) address);
 }
 
 static uint32_t rotate(uint8_t motor, int32_t velocity)
