+++ boards/TMC2300_eval.c
@@ -115,14 +115,14 @@ uint8_t tmc2300_CRC8(uint8_t *data, size_t length)
 }
 // <= CRC wrapper
 
-void tmc2300_writeRegister(uint8_t motor, uint8_t address, int32_t value)
+void tmc2300_writeRegister(uint8_t motor, uint16_t address, int32_t value)
 {
-	tmc2300_writeInt(motorToIC(motor), address, value);
+	tmc2300_writeInt(motorToIC(motor), (uint8_t) address, value);
 }
 
-void tmc2300_readRegister(uint8_t motor, uint8_t address, int32_t *value)
+void tmc2300_readRegister(uint8_t motor, uint16_t address, int32_t *value)
 {
-	*value = tmc2300_readInt(motorToIC(motor), address);
+	*value = tmc2300_readInt(motorToIC(motor), (uint8_t) address);
 }
 
 static uint32_t rotate(uint8_t motor, int32_t velocity)
