+++ boards/TMC7300_eval.c
@@ -95,14 +95,14 @@ uint8_t tmc7300_CRC8(uint8_t *data, size_t length)
 }
 // <= CRC wrapper
 
-void tmc7300_writeRegister(uint8_t motor, uint8_t address, int32_t value)
+void tmc7300_writeRegister(uint8_t motor, uint16_t address, int32_t value)
 {
-	tmc7300_writeInt(motorToIC(motor), address, value);
+	tmc7300_writeInt(motorToIC(motor), (uint8_t) address, value);
 }
 
-void tmc7300_readRegister(uint8_t motor, uint8_t address, int32_t *value)
+void tmc7300_readRegister(uint8_t motor, uint16_t address, int32_t *value)
 {
-	*value = tmc7300_readInt(motorToIC(motor), address);
+	*value = tmc7300_readInt(motorToIC(motor), (uint8_t) address);
 }
 
 static uint32_t rotate(uint8_t motor, int32_t velocity)
